Simplify pre-compiled regular expressions for sigils
I discovered `tagref` today, when looking for a system inspired
from GHC notes. I was delighted and tried it on a code-base where I
have been writing notes, but it didn't work! Let me describe the
problem:

* Setup: Here is my sysinfo, as published by `neofetch`:
     OS: macOS 14.3.1 23D60 arm64
     Host: MacBookAir10,1
     Shell: fish 3.7.0
     Terminal: tmux
     CPU: Apple M1

On this system, `tagref` did not recognize any tags in my code, even
when I had tags.
```
─> tagref
0 tags, 0 tag references, 0 file references, and 0 directory references validated in 464 files.
```

The problem seems to be in the capture group used in the regular
expressions, which I verified using `rg` and `gsed`:
```
─> rg "\[\s*tag\s*:.*\s*\]"
components/whatsapp/src/ai/salher/whatsapp/core.clj
267:;; ## [tag:GIFs need to be converted to video/mp4 format]

components/chapters/src/ai/salher/chapters/core.clj
173:;;; ## [tag:Rules for deciding when we should update a chapter record]
185:;;; ## [tag:Rules to mark a chapter as DONE for now]
...
```
The current capture group does not support spaces in the identifier
name!
```
─> cat components/whatsapp/src/ai/salher/whatsapp/core.clj | gsed -En '/\[\s*tag\s*:\s*([^\]\s]*)\]/p'

─> cat components/whatsapp/src/ai/salher/whatsapp/core.clj | gsed -En '/\[\s*tag\s*:\s*.*\]/p'
;; ## [tag:GIFs need to be converted to video/mp4 format]

─> cat components/whatsapp/src/ai/salher/whatsapp/core.clj | gsed -En 's/.*\[\s*tag\s*:\s*(.*)\s*\].*/\1/p'
GIFs need to be converted to video/mp4 format
```

On seeing this, I tried to understand what the expressions are trying
to do, and add support for spaces. My current understanding is as follows:

1. Look for a starting square-bracket `[`
2. There can be 0 or more spaces between the square bracket and the
sigil, ignore them.
3. Match the sigil
4. There can be 0 or more spaces between the sigil and `:` token,
ignore them
5. There can be 0 or more spaces between the `:` token and the actual
identifier, ignore them
6. Capture the identifier.
7. There can be 0 or more spaces between the end of the identifier and
the closing square-bracket, ignore them.
8. Stop at the closing square-bracket `]`

With this in mind, I changed the capture group to a lazy-capture (from
`([^\\]\\s]*)` to `(\\w.*?)`. This ensures that everything is captured
until the closing square bracket (ignoring all empty space immediately
before the closing bracket). The only constraint it imposes is that
the first character of the identifier should be a letter, digit or
underscore (via the `\w`). This is to ensure that `[tag:]` is not
treated as a valid tag.

After this change, `tagref` is now working correctly on my local
system. I ran `cargo tests` (all passing) and also tested on my own
project, which now shows:
```
─> tagref
21 tags, 24 tag references, 0 file references, and 0 directory references validated in 464 files.
```

Update: based on feedback received in the PR #218, I have changed the
capture group from `(\\w.*?)` to `(.+?)`, to support identifiers
like `[file:.gitignore]`. This means a tag needs to be at least one
character which is not a line-break.